feat: draggable lesson update

+++ app/Http/Controllers/Instructor/CourseChapterController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\CourseChapter\StoreRequest;
 use App\Http\Requests\CourseChapter\UpdateRequest;
 use App\Models\Course;
 use App\Models\CourseChapter;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 
 class CourseChapterController extends Controller
@@ -73,7 +74,36 @@ class CourseChapterController extends Controller
         if ($chapter->course_id != $course->id) {
             return response()->json(['message' => 'Unauthorized'], 403);
         }
+        if ($chapter->lessons()->exists()) {
+            return response()->json(['message' => 'Chapter has lessons and cannot be deleted'], 403);
+        }
         $chapter->delete();
         return response()->json(['message' => 'Chapter deleted successfully'], 200);
+    }
+
+    public function orderLessons(Request $request, Course $course, CourseChapter $chapter)
+    {
+        $request->validate([
+            'lessons_ids' => 'required|array',
+            'lessons_ids.*' => 'exists:lessons,id',
+        ]);
+
+        $instructor_id = Auth::user()->id;
+        if ($chapter->instructor_id != $instructor_id) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+        if ($chapter->course_id != $course->id) {
+            return response()->json(['message' => 'Unauthorized'], 403);
+        }
+        // Logic to order lessons
+        $lessons_ids = $request->input('lessons_ids');
+        foreach ($lessons_ids as $order => $lesson_id) {
+            $lesson = $chapter->lessons()->find($lesson_id);
+            if ($lesson) {
+                $lesson->order = $order;
+                $lesson->save();
+            }
+        }
+        return response()->json(['message' => 'Lessons ordered successfully'], 200);
     }
 }
